requirements.txt
Flask==3.1.1
gunicorn==23.0.0
pypdf==4.3.1
PyMuPDF==1.24.14
pytesseract==0.3.13
Pillow==11.2.1
